refactor(controller/wavs): extract waveform rendering helpers
